+++ iTunesAppJisung/Presentation/Shared/View/OpenButton.swift
@@ -12,6 +12,8 @@ final class OpenButton: UIButton {
     }
 
     private func configureUI() {
+        addTarget(self, action: #selector(didButtonTapped), for: .touchUpInside)
+
         setTitle("열기", for: .normal)
         setTitleColor(.systemBlue, for: .normal)
         titleLabel?.font = .boldSystemFont(ofSize: 16)
@@ -20,4 +22,7 @@ final class OpenButton: UIButton {
         layer.cornerRadius = 16
         clipsToBounds = true
     }
+
+    @objc
+    private func didButtonTapped() {}
 }
